Project Bits/requirements.txt
requests
geocoder
win10toast
tk
Pillow